BUG: Handling of signed values in mul.wide now determines the signs of the inputs from the instruction type, not the register types.
Former-commit-id: 2b60e62ea92754e7dbe3880a0c27d0417179c01c

+++ branches/ocelot-ptx-2.1/ocelot/analysis/implementation/StructuralAnalysis.cpp
@@ -310,7 +310,7 @@ namespace analysis {
       if (debug) {
         dumpCTNode(p);
         dumpUnstructuredBR();
-        printf("%ld\n", N.size());
+        printf("%ld\n", (long int)N.size());
         std::cerr << "********************\n";
       }
